+++ comb_spec_searcher/comb_spec_searcher.py
@@ -26,7 +26,7 @@ from logzero import logger
 from comb_spec_searcher.typing import CombinatorialClassType, CSSstrategy
 
 from .class_db import ClassDB
-from .class_queue import DefaultQueue
+from .class_queue import CSSQueue, DefaultQueue
 from .exception import (
     ExceededMaxtimeError,
     InvalidOperationError,
@@ -70,6 +70,8 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         strategy_pack: StrategyPack,
         *,
         ruledb: Optional[RuleDBAbstract] = None,
+        classdb: Optional[ClassDB[CombinatorialClassType]] = None,
+        classqueue: Optional[CSSQueue] = None,
         expand_verified: bool = False,
         debug: bool = False,
     ):
@@ -97,8 +99,14 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         self.func_calls: Dict[str, int] = defaultdict(int)
         self.func_yield: Dict[str, int] = defaultdict(int)
 
-        self.classdb = ClassDB[CombinatorialClassType](type(start_class))
-        self.classqueue = DefaultQueue(strategy_pack)
+        self.classdb = (
+            classdb
+            if classdb is not None
+            else ClassDB[CombinatorialClassType](type(start_class))
+        )
+        self.classqueue = (
+            DefaultQueue(strategy_pack) if classqueue is None else classqueue
+        )
         self.ruledb: RuleDBAbstract = ruledb if ruledb is not None else RuleDB()
         self.ruledb.link_searcher(self)
 
@@ -400,7 +408,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             + tabulate.tabulate(table, headers=headers, colalign=colalign).replace(
                 "\n", "\n    "
             )
-            + "\n"
+            + "\n\n"
         )
 
     def _mem_status(self, elaborate: bool) -> str: